plugins: lfa: Adds reaction to flow up/down events
This adds functionality in the LFA policy to react to N-1 flow up and
down events. When an N-1 flow goes down, a LFA is searched for N flows
that have the N-1 flow as current next hop. When a flow goes up, the
behaviour depends on a module parameter. If no reaction is required,
the event is just logged. If reaction is required, the N flows are
checked. The primary hop is re-set as next hop if it was down. If it
was down and the N flow's current next hop differs from its primary
next hop, a LFA is searched for, since a new one may have become
available.